fix: air quality none (#417)
Air quality for purifiers that don't support it was being converted from
a value of None to a string of "none" which made it show up in the HA
interface. I also noticed a few other items where if None passed in it
isn't set. Not aware of that causing a bug but should be fixed. Also
removed -1 as a value as we decided to use "None" for unsupported.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>